gateway-service/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
grpcio>=1.59.0
grpcio-tools>=1.59.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-multipart>=0.0.6
pytest>=7.4.3
httpx>=0.25.0
psutil>=5.9.0